+++ CHANGELOG.md
@@ -20,6 +20,7 @@ require depending on default function and bootloader code
 - Support for Elixir `Enum` functions: `split_with`, `join`, `map_join`, `into`, `reverse`,
 `to_list`
 - Support for Elixir `MapSet` module
+- Support for Elixir `Range` module
 
 ## [0.6.3] - 20-07-2024